+++ v4l2py/device.py
@@ -17,7 +17,6 @@ import mmap
 import os
 import pathlib
 import typing
-import configparser
 from io import IOBase
 from collections import UserDict
 
@@ -1268,143 +1267,6 @@ class LegacyControl(BaseNumericControl):
 
     def decrease(self, steps: int = 1):
         self.value -= steps * self.step
-
-
-class ConfigManager:
-    def __init__(self, device: Device):
-        self.device = device
-        self.log = log.getChild(f"config({device.index})")
-        self.config = None
-        self.filename = None
-
-    @property
-    def has_config(self) -> bool:
-        return isinstance(self.config, configparser.ConfigParser) and self.config.sections()
-
-    @property
-    def config_loaded(self) -> bool:
-        return self.filename is not None
-
-    def reset(self) -> None:
-        self.config = configparser.ConfigParser()
-        self.filename = None
-
-    def acquire(self) -> None:
-        self.log.info(f"acquiring configuration from {self.device.filename}")
-        if not self.has_config:
-            self.reset()
-
-        self.config["device"] = {
-            "driver": str(self.device.info.driver),
-            "card": str(self.device.info.card),
-            "bus_info": str(self.device.info.bus_info),
-            "version": str(self.device.info.version)
-        }
-        self.config["controls"] = {}
-        for c in self.device.controls.values():
-            self.config["controls"][c.config_name] = str(c.value)
-        self.log.info("configuration successfully acquired")
-
-    def save(self, filename) -> None:
-        self.log.info(f"writing configuration to {filename}")
-        if isinstance(filename, pathlib.Path):
-            pass
-        elif isinstance(filename, str):
-            filename = pathlib.Path(filename)
-        else:
-            raise TypeError(f"filename expected to be str or pathlib.Path, not {filename.__class__.__name__}")
-
-        if self.device.closed:
-            raise RuntimeError(f"{self.device}: must be opened to save configuration")
-        if not self.config or not self.config.sections():
-            self.acquire()
-
-        with filename.open(mode="wt") as configfile:
-            self.config.write(configfile)
-        self.log.info(f"configuration written to {filename.resolve()}")
-
-    def load(self, filename) -> None:
-        self.log.info(f"reading configuration from {filename}")
-
-        if isinstance(filename, pathlib.Path):
-            pass
-        elif isinstance(filename, str):
-            filename = pathlib.Path(filename)
-        else:
-            raise TypeError(f"filename expected to be str or pathlib.Path, not {filename.__class__.__name__}")
-
-        if not (filename.exists() and filename.is_file()):
-            raise RuntimeError(f"{filename} must be an existing file")
-        if self.device.closed:
-            raise RuntimeError(f"{self.device}: must be opened to load configuration")
-
-        self.reset()
-        res = self.config.read((filename,))
-        if not res:
-            raise RuntimeError(f"Failed to read configuration from {filename}")
-        else:
-            filename = pathlib.Path(res[0])
-            self.filename = filename.resolve()
-        self.log.info(f"configuration read from {self.filename}")
-
-    def validate(self, pedantic: bool = False) -> None:
-        self.log.info("validating configuration")
-        if not self.config_loaded:
-            raise RuntimeError("Load configuration first")
-        for section in ("controls",):
-            if not self.config.has_section(section):
-                raise RuntimeError(f"Mandatory section '{section}' is missing")
-        controls = self.device.controls.named_keys()
-        for ctrl, _ in self.config.items("controls"):
-            if ctrl not in controls:
-                raise RuntimeError(f"{self.device.filename} has no control named {ctrl}")
-
-        if pedantic:
-            if not self.config.has_section("device"):
-                raise RuntimeError("Section 'device' is missing")
-            for (option, have) in (
-                ("card", str(self.device.info.card)),
-                ("driver", str(self.device.info.driver)),
-                ("version", str(self.device.info.version)),
-            ):
-                want = self.config["device"][option]
-                if not (want == have):
-                    raise RuntimeError(f"{option.title()} mismatch: want '{want}', have '{have}'")
-        self.log.info("configuration validated")
-
-    def apply(self, cycles: int = 2) -> None:
-        self.log.info("applying configuration")
-        if not self.config_loaded:
-            raise RuntimeError("Load configuration first")
-
-        for cycle in range(1, cycles + 1):
-            for ctrl, value in self.config.items("controls"):
-                what = f"#{cycle}/{cycles} {ctrl}"
-                if self.device.controls[ctrl].is_writeable:
-                    if not self.device.controls[ctrl].is_flagged_write_only:
-                        cur = f"{self.device.controls[ctrl].value}"
-                    else:
-                        cur = "<not_readable>"
-                    self.log.debug(f"{what} {cur} => {value}")
-                    self.device.controls[ctrl].value = value
-                else:
-                    self.log.debug(f"{what} skipped (not writeable)")
-        self.log.info("configuration applied")
-
-    def verify(self) -> None:
-        self.log.info("verifying device configuration")
-        if not self.config_loaded:
-            raise RuntimeError("Load configuration first")
-
-        for ctrl, value in self.config.items("controls"):
-            if not self.device.controls[ctrl].is_flagged_write_only:
-                cur = str(self.device.controls[ctrl].value)
-                self.log.debug(f"{ctrl}: want {value}, have {cur}")
-                if not cur.lower() == value.lower():
-                    raise RuntimeError(f"{ctrl} should be {value}, but is {cur}")
-            else:
-                self.log.debug(f"{ctrl} skipped (not readable)")
-        self.log.info("device is configured correctly")
 
 
 class DeviceHelper:
